+++ packages/core/CHANGELOG.md
@@ -1,5 +1,15 @@
 # @ayazmo/core
 
+## 0.5.1
+
+### Patch Changes
+
+- cache entities
+
+- Updated dependencies []:
+  - @ayazmo/utils@0.5.1
+  - @ayazmo/types@0.5.1
+
 ## 0.5.0
 
 ### Minor Changes