+++ packages/types/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @ayazmo/types
 
+## 0.5.1
+
+### Patch Changes
+
+- cache entities
+
 ## 0.5.0
 
 ### Minor Changes