+++ packages/utils/CHANGELOG.md
@@ -1,5 +1,14 @@
 # @ayazmo/utils
 
+## 0.5.1
+
+### Patch Changes
+
+- cache entities
+
+- Updated dependencies []:
+  - @ayazmo/types@0.5.1
+
 ## 0.5.0
 
 ### Minor Changes